+++ services/BlogService.js
@@ -2,7 +2,7 @@ var MarkdownIt = require('markdown-it')()
 
 const Service = require('./Service')
 const Database = require('../sql')
-const config = require('../config')
+const twilioUtil = require('../util/twilio')
 
 const Blog = Database.Blog
 const User = Database.User
@@ -17,32 +17,6 @@ function requireAuthenticated(loggedInUser) {
             message: 'Unauthorized',
             status: 401,
         }
-    }
-}
-
-async function sendNotifications(author) {
-    console.log('Entered Send Notification')
-    const client = require('twilio')(
-        config.TWILIO_CLIENT_ID,
-        config.TWILIO_CLIENT_SECRET
-    )
-    console.log('Created Client')
-    var followers = await Follow.findAll({ where: { followee: author.uuid } })
-    for (var i = 0; i < followers.length; i++) {
-        var followerInfo = await User.findAll({
-            where: { uuid: followers[i].follower },
-        })
-        client.messages
-            .create({
-                body:
-                    'Hello, a new blog post from ' +
-                    author.name +
-                    ' has been posted and is ready for you to read.',
-                to: followerInfo[0].phone, // Text this number
-                from: config.TWILIO_PHONE_NUMBER, // From a valid Twilio number
-            })
-            .then((message) => console.log(message.sid))
-        console.log('Created Message for: ', followerInfo[0].name)
     }
 }
 
@@ -248,7 +222,7 @@ const apiBlogsUuidPUT = ({ uuid, body }, loggedInUser) =>
                     limit: 1,
                 })
                 console.log('Author Results:' + JSON.stringify(author))
-                sendNotifications(author[0])
+                twilioUtil.sendNotifications(author[0])
             }
 
             await blog.save()
