+++ tests/server/integration/blog.test.js
@@ -41,9 +41,3 @@ afterEach(async (done) => {
     server = null
     done()
 })
-
-afterAll(async (done) => {
-    // close db after completion
-    await db.close()
-    done()
-})
