+++ tests/server/unit/services/apiBlogsUuidPUT.test.js
@@ -10,8 +10,13 @@ import 'regenerator-runtime/runtime'
 
 const dbHelper = require('../../dbHelper')
 const db = require('../../../../sql')
+const twilioUtil = require('../../../../util/twilio')
 
 const delay = (ms) => new Promise((res) => setTimeout(res, ms))
+
+beforeAll(() => {
+    twilioUtil.setAPIEnabled(false)
+})
 
 describe('blog post Uuid PUT handler tests', () => {
     test('invalid blog uuid returns reject error', async () => {
@@ -413,10 +418,4 @@ describe('blog post Uuid PUT handler tests', () => {
         expect(typeof data2.payload.authorName).toBe('string')
         expect(data2.payload.authorName).toBe('John Doe')
     })
-})
-
-afterAll(async (done) => {
-    // close db after completion
-    await db.close()
-    done()
 })
